Move amount input into component
